+++ app/src/main/java/com/example/fap/data/Entities/Payment.kt
@@ -1,0 +1,42 @@
+package com.example.fap.data.entities
+
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
+import java.util.Date
+
+@Entity(
+    tableName = "Payment",
+    foreignKeys = [
+        ForeignKey(
+            entity = User::class,
+            parentColumns = ["id"],
+            childColumns = ["userId"],
+            onDelete = ForeignKey.CASCADE
+        ),
+        ForeignKey(
+            entity = Wallet::class,
+            parentColumns = ["id"],
+            childColumns = ["walletId"],
+            onDelete = ForeignKey.CASCADE
+        ),
+        ForeignKey(
+            entity = Category::class,
+            parentColumns = ["id"],
+            childColumns = ["categoryId"],
+            onDelete = ForeignKey.CASCADE
+        )
+    ]
+)
+data class Payment(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
+    val userId: String,
+    val walletId: Int,
+    val title: String,
+    val description: String,
+    val price: Double,
+    val date: Date,
+    val isPayment: Boolean,
+    val categoryId: Int
+)
