+++ app/src/main/java/com/example/fap/data/Entities/SavingGoal.kt
@@ -1,0 +1,38 @@
+package com.example.fap.data.entities
+
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
+import java.util.Date
+
+@Entity(
+    tableName = "SavingsGoal",
+    foreignKeys = [
+        ForeignKey(
+            entity = User::class,
+            parentColumns = ["id"],
+            childColumns = ["userId"],
+            onDelete = ForeignKey.CASCADE
+        ),
+        ForeignKey(
+            entity = Wallet::class,
+            parentColumns = ["id"],
+            childColumns = ["walletId"],
+            onDelete = ForeignKey.CASCADE
+        )
+    ]
+)
+
+data class SavingsGoal(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
+    val userId: String,
+    val walletId: Int,
+    val title: String,
+    val description: String,
+    val startDate: Date,
+    val endDate: Date,
+    val amountPerMonth: Double,
+    val endAmount: Double,
+    val startAmount: Double
+)
