+++ app/src/main/java/com/example/fap/data/Entities/Stock.kt
@@ -1,0 +1,32 @@
+package com.example.fap.data.entities
+
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
+
+@Entity(
+    tableName = "Stock",
+    foreignKeys = [
+        ForeignKey(
+            entity = User::class,
+            parentColumns = ["id"],
+            childColumns = ["userId"],
+            onDelete = ForeignKey.CASCADE
+        ),
+        ForeignKey(
+            entity = Wallet::class,
+            parentColumns = ["id"],
+            childColumns = ["walletId"],
+            onDelete = ForeignKey.CASCADE
+        )
+    ]
+)
+data class Stock(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
+    val userId: String,
+    val walletId: Int,
+    val name: String,
+    val amount: Int,
+    val price: Double
+)
