+++ app/src/main/java/com/example/fap/data/Entities/User.kt
@@ -1,0 +1,10 @@
+package com.example.fap.data.entities
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "User")
+data class User(
+    @PrimaryKey val id: String,
+    val name: String? = null
+)
