+++ app/src/main/java/com/example/fap/data/Entities/Wallet.kt
@@ -1,0 +1,21 @@
+package com.example.fap.data.entities
+
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
+import com.example.fap.data.entities.User
+
+@Entity(tableName = "Wallet",
+    foreignKeys = [ForeignKey(
+        entity = User::class,
+        parentColumns = ["id"],
+        childColumns = ["userId"],
+        onDelete = ForeignKey.CASCADE
+    )]
+)
+data class Wallet(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
+    val userId: String,
+    val name: String
+)
