+++ app/src/main/java/com/example/fap/data/dao/Category.kt
@@ -1,0 +1,16 @@
+package com.example.fap.data.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.Query
+import com.example.fap.data.entities.Category
+
+@Dao
+interface FapDaoCategory {
+
+    @Query("SELECT * FROM Category WHERE userId = :userId")
+    suspend fun getCategories(userId: Int): List<Category>
+
+    @Insert
+    suspend fun insertCategory(category: Category): Long
+}
