+++ app/src/main/java/com/example/fap/data/dao/SavingGoal.kt
@@ -1,0 +1,16 @@
+package com.example.fap.data.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.Query
+import com.example.fap.data.entities.SavingsGoal
+
+@Dao
+interface FapDaoSavingGoal {
+
+    @Query("SELECT * FROM SavingsGoal WHERE userId = :userId")
+    suspend fun getSavingsGoals(userId: Int): List<SavingsGoal>
+
+    @Insert
+    suspend fun insertSavingsGoal(savingsGoal: SavingsGoal): Long
+}
