+++ app/src/main/java/com/example/fap/data/dao/User.kt
@@ -1,0 +1,17 @@
+package com.example.fap.data.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.Query
+import com.example.fap.data.entities.User
+
+@Dao
+interface FapDaoUser {
+
+    @Insert
+    suspend fun insertUser(user: User): Long
+
+    @Query("SELECT * FROM user")
+    suspend fun getUsers(): List<User>
+
+}
